Move some code at a better place.
